Time: 0 ms (100.00%), Space: 5.9 MB (22.94%) - LeetHub

+++ 0374-guess-number-higher-or-lower/0374-guess-number-higher-or-lower.cpp
@@ -11,12 +11,12 @@ class Solution {
 public:
     int guessNumber(int n) {
         
-       long long l=1,r=n;
+       int l=1,r=n;
         
         long long m,res;
         
         while(l<=r){
-            m=(l+r)/2;
+            m=((long long)l+r)/2;
             
             if(guess(m)==0){
                 res=m;
